Model changes to support the dashboard requirements.

+++ src/main/java/org/geppetto/core/data/IGeppettoDataManager.java
@@ -49,7 +49,7 @@ public interface IGeppettoDataManager
 	
 	IUser getUserByLogin(String login);
 	
-	IGeppettoProject getGeppettoProjectById(int id);
+	IGeppettoProject getGeppettoProjectById(long id);
 
 	List<? extends IGeppettoProject> getAllGeppettoProjects();
